refactor: Update group migrations and models
This commit updates the group migrations and models to improve the structure and functionality of the groups feature.

- In the 'create_groups_table' migration, the 'groupID' column is changed to use the 'id' method for primary key generation. The 'groupName' column is also modified to remove the length constraint.
- In the 'create_users_groups_table' migration, the 'groupID' column is changed to use the 'unsignedBigInteger' type for better compatibility with foreign key references.

These changes ensure better consistency and compatibility in the database schema for groups.

+++ database/migrations/2024_06_10_123842_create_users_groups_table.php
@@ -12,7 +12,7 @@ return new class extends Migration
     public function up(): void
     {
         Schema::create('users_groups', function (Blueprint $table) {
-            $table->integer('groupID');
+            $table->unsignedBigInteger('groupID');
             $table->foreign('groupID')->references('groupID')->on('groups')->onUpdate('cascade')->onDelete('cascade');
             $table->foreignId('id')->constrained(
                 table: 'users', indexName:'id4', column:'id'
